# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Fleet Feast API
  description: |
    RESTful API for Fleet Feast food truck marketplace platform.

    ## Overview
    Fleet Feast connects customers with food truck vendors for private events. The API supports:
    - User authentication and authorization with role-based access control
    - Vendor application, approval, and profile management
    - Food truck search and discovery with filters
    - Booking requests, acceptance, and cancellation
    - Escrow payments via Stripe Connect
    - In-app messaging with anti-circumvention monitoring
    - Review and rating system
    - Admin moderation and analytics

    ## Authentication
    All protected endpoints require authentication via JWT token stored in HTTP-only secure cookie.
    The token is set after successful login via `/api/auth/login` endpoint.

    ## Rate Limiting
    - **Authenticated users**: 100 requests per minute
    - **Unauthenticated users**: 20 requests per minute

    Rate limit information is included in response headers:
    - `X-RateLimit-Limit`: Total requests allowed per window
    - `X-RateLimit-Remaining`: Requests remaining in current window
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

    ## Pagination
    List endpoints use cursor-based pagination for performance:
    - `cursor`: Opaque cursor string for next page
    - `limit`: Number of items per page (default: 20, max: 100)

    ## Error Handling
...

  version: 1.0.0
  contact:
...
  license:
...
servers:
...
paths:
...
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register new user account
      description: Create a new customer or vendor account. Email verification required before full access.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          userId:
                            type: string
                            format: uuid
                          email:
                            type: string
                            format: email
                          role:
                            $ref: '#/components/schemas/UserRole'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
...
  /api/auth/verify-email:
    post:
      tags: [Authentication]
      summary: Verify email address
      description: Verify user email with token sent to email address
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  description: Email verification token
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/auth/reset-password:
    post:
      tags: [Authentication]
      summary: Request password reset
      description: Send password reset email to user
      operationId: requestPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Password reset email sent (always returns 200 to prevent email enumeration)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token stored in HTTP-only secure cookie
  schemas:
...
    ApiResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
...
    UserRole:
      type: string
      enum: [CUSTOMER, VENDOR, ADMIN]
...
    RegisterRequest:
      type: object
      required: [email, password, role]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          $ref: '#/components/schemas/UserRole'
...
  responses:
    BadRequest:
      description: Bad request - invalid parameters or request body
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        enum: [VALIDATION_ERROR, INVALID_INPUT]
                      message:
                        type: string
                      details:
                        type: object
...
    NotFound:
      description: Not found - resource does not exist
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        enum: [NOT_FOUND]
                      message:
                        type: string
    Conflict:
      description: Conflict - resource already exists or operation conflicts with current state
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        enum: [CONFLICT, DUPLICATE_RESOURCE, INVALID_STATE_TRANSITION]
                      message:
                        type: string
    RateLimitExceeded:
      description: Rate limit exceeded - too many requests
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        enum: [RATE_LIMIT_EXCEEDED]
                      message:
                        type: string
                      details:
                        type: object
                        properties:
                          retryAfter:
                            type: integer
                            description: Seconds until rate limit resets